# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vyndi API
  description: |
    API documentation for Vyndi - AI-powered business offer generation platform.

    This API provides endpoints for:
    - Authentication and session management
    - Offer generation and management
    - PDF generation and export
    - Team collaboration
    - User profile and settings

    ## Authentication

    All authenticated endpoints require a valid session token in cookies:
    - `propono_at` - Access token (HTTP-only cookie)
    - `propono_rt` - Refresh token (HTTP-only cookie)
    - `XSRF-TOKEN` - CSRF token (sent in header as `x-csrf-token` for state-changing operations)

    ## Rate Limiting

    All API endpoints implement rate limiting. Response headers include:
    - `X-RateLimit-Limit` - Maximum requests allowed in the time window
    - `X-RateLimit-Remaining` - Remaining requests in the current window
    - `X-RateLimit-Reset` - ISO timestamp when the rate limit resets
    - `Retry-After` - Seconds to wait before retrying (on 429 responses)

    ## Error Responses

    All errors follow a consistent format with an `error` message, optional `requestId` for tracking, and optional `issues` for validation errors.
  version: 1.0.0
  contact:
    name: Vyndi Support
    email: support@vyndi.com
  license:
    name: Proprietary
servers:
  - url: https://vyndi.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/auth/session:
    get:
      tags:
        - Authentication
      summary: Get current session
      description: Returns the current user session information
      responses:
        '200':
          description: Session information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  session:
                    $ref: '#/components/schemas/Session'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/offers/{offerId}/share:
    get:
      tags:
        - Offers
      summary: List share links for an offer
      description: |
        Retrieves all share links for a specific offer.
        Requires a paid plan (standard or pro).
      security:
        - cookieAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share links retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  shares:
                    type: array
                    items:
                      $ref: '#/components/schemas/OfferShare'
        '402':
          description: Payment required - feature requires paid plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not the offer owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Offers
      summary: Create share link for an offer
      description: |
        Creates a new shareable link for an offer.
        Requires a paid plan (standard or pro).
      security:
        - cookieAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresAt:
                  type: string
                  format: date-time
                  description: Optional expiration date for the share link
                customerEmail:
                  type: string
                  format: email
                  description: Optional customer email for tracking
                customerName:
                  type: string
                  maxLength: 200
                  description: Optional customer name for tracking
      responses:
        '200':
          description: Share link created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  shareId:
                    type: string
                    format: uuid
                  shareUrl:
                    type: string
                    format: uri
                  token:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    nullable: true
                  createdAt:
                    type: string
                    format: date-time
        '402':
          description: Payment required - feature requires paid plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not the offer owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: propono_at
      description: Session access token stored in HTTP-only cookie
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
    Session:
      type: object
      properties:
        access_token:
          type: string
        expires_at:
          type: string
          format: date-time
...
    OfferShare:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
          description: Share token for the shareable URL
        shareUrl:
          type: string
          format: uri
          description: Full shareable URL
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Expiration date of the share link
        isActive:
          type: boolean
          description: Whether the share link is currently active
        accessCount:
          type: integer
          description: Number of times the share link has been accessed
        lastAccessedAt:
          type: string
          format: date-time
          nullable: true
          description: Last access timestamp
        customerEmail:
          type: string
          format: email
          nullable: true
          description: Associated customer email
        customerName:
          type: string
          nullable: true
          description: Associated customer name
        createdAt:
          type: string
          format: date-time
          description: Share link creation timestamp
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        requestId:
          type: string
          format: uuid
          description: Request ID for tracking
        issues:
          type: object
          nullable: true
          description: Validation errors (if applicable)
          properties:
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            formErrors:
              type: array
              items:
                type: string
        requiresUpgrade:
          type: boolean
          nullable: true
          description: Whether the error is due to requiring a paid plan
        feature:
          type: string
          nullable: true
          description: Feature name that requires upgrade